+++ lib/karafka/routing/features/active_job/contract.rb
@@ -10,7 +10,7 @@ module Karafka
           configure do |config|
             config.error_messages = YAML.safe_load(
               File.read(
-                File.join(Karafka.gem_root, 'config', 'errors.yml')
+                File.join(Karafka.gem_root, 'config', 'locales', 'errors.yml')
               )
             ).fetch('en').fetch('validations').fetch('topic')
           end