+++ spec/lib/karafka/contracts/base_spec.rb
@@ -6,7 +6,7 @@ RSpec.describe_current do
       configure do |config|
         config.error_messages = YAML.safe_load(
           File.read(
-            File.join(Karafka.gem_root, 'config', 'errors.yml')
+            File.join(Karafka.gem_root, 'config', 'locales', 'errors.yml')
           )
         ).fetch('en').fetch('validations').fetch('test')
       end